# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sisCotacao — API",
    "description": "Contrato dos endpoints principais do ERP (fase 1). Autenticação por Bearer token emitido em /api/login (TTL 7 dias, revogável por logout ou troca de senha). Portal do fornecedor usa token próprio na URL e está fora deste contrato.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem (nginx proxy /api)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/precos/configuracao": {
      "get": {
        "tags": ["precos"],
        "summary": "Consulta as premissas persistidas da formação de preço",
        "responses": {"200": {"description": "Configuração de precificação"}}
      },
      "put": {
        "tags": ["precos"],
        "summary": "Atualiza faturamento, despesas, tributos, atividade e cenário",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"type": "object"}}}
        },
        "responses": {
          "200": {"description": "Configuração atualizada"},
          "400": {"description": "Premissa inválida"}
        }
      }
    },
    "/api/precos/calcular/{produto_id}": {
      "get": {
        "tags": ["precos"],
        "summary": "Calcula preço pela metodologia do divisor e devolve memória de cálculo",
        "parameters": [
          {"name": "produto_id", "in": "path", "required": true, "schema": {"type": "integer"}},
          {"name": "margem", "in": "query", "schema": {"type": "number"}},
          {"name": "embalagem_unitaria", "in": "query", "schema": {"type": "number", "minimum": 0}},
          {"name": "frete_unitario", "in": "query", "schema": {"type": "number", "minimum": 0}},
          {"name": "frete_pct", "in": "query", "schema": {"type": "number", "minimum": 0, "maximum": 100}},
          {"name": "cartao_pct", "in": "query", "schema": {"type": "number", "minimum": 0, "maximum": 100}},
          {"name": "impostos_pct", "in": "query", "schema": {"type": "number", "minimum": 0, "maximum": 100}},
          {"name": "cenario_tributario", "in": "query", "schema": {"type": "string", "enum": ["atual", "reforma"]}}
        ],
        "responses": {
          "200": {"description": "Preço e memória de cálculo"},
          "400": {"description": "Parâmetro inválido"}
        }
      }
    },
    "/api/precos/regras/{produto_id}": {
      "get": {
        "tags": [
          "precos"
        ],
        "summary": "Lista regras de pre?o do produto (MDM-007)",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Regras ativas"
          }
        }
      },
      "post": {
        "tags": [
          "precos"
        ],
        "summary": "Cria regra de pre?o (prioridade, contexto canal/cliente/segmento/quantidade, vig?ncia, pre?o ou desconto, margem m?nima)",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prioridade": {
                    "type": "integer",
                    "default": 10
                  },
                  "canal": {
                    "type": "string",
                    "nullable": true
                  },
                  "cliente_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "segmento": {
                    "type": "string",
                    "nullable": true
                  },
                  "quantidade_min": {
                    "type": "number",
                    "nullable": true
                  },
                  "preco": {
                    "type": "number",
                    "nullable": true
                  },
                  "desconto_pct": {
                    "type": "number",
                    "nullable": true
                  },
                  "margem_minima_pct": {
                    "type": "number",
                    "nullable": true
                  },
                  "vigencia_inicio": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "vigencia_fim": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "motivo": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Regra salva"
          },
          "400": {
            "description": "Regra inv?lida"
          }
        }
      }
    },
    "/api/precos/regras/{produto_id}/{regra_id}": {
      "delete": {
        "tags": [
          "precos"
        ],
        "summary": "Remove (desativa) uma regra de pre?o do produto",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removida"
          },
          "404": {
            "description": "N?o encontrada"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "HMAC (base64.payload.signature), TTL 7 dias"
      }
    },
...
  }
}